+++ examples/CMakeLists.txt
@@ -2,3 +2,4 @@ cmake_minimum_required(VERSION 2.8)
 project(examples)
 
 add_subdirectory(aditof-demo)
+add_subdirectory(first-frame)
