+++ examples/first-frame/CMakeLists.txt
@@ -1,0 +1,8 @@
+cmake_minimum_required(VERSION 2.8)
+project(first-frame)
+
+add_executable(${PROJECT_NAME} main.cpp)
+
+set_target_properties(${PROJECT_NAME} PROPERTIES CXX_STANDARD 11)
+
+target_link_libraries(${PROJECT_NAME} PRIVATE aditof)
